Fix Compose API compatibility issues
- Add .toFloat() conversion in ComposeRenderer for IsoColor to Color
  (fixes type mismatch: IsoColor uses Double, Compose Color needs Float)
- Add missing content lambda to IsometricCanvas in AnimatedSample

Resolves compilation errors with Compose 1.5.0 type requirements.

+++ isometric-compose/src/main/kotlin/io/fabianterhorst/isometric/compose/ComposeRenderer.kt
@@ -53,10 +53,10 @@ object ComposeRenderer {
      */
     fun IsoColor.toComposeColor(): Color {
         return Color(
-            red = (r / 255f).coerceIn(0f, 1f),
-            green = (g / 255f).coerceIn(0f, 1f),
-            blue = (b / 255f).coerceIn(0f, 1f),
-            alpha = (a / 255f).coerceIn(0f, 1f)
+            red = (r.toFloat() / 255f).coerceIn(0f, 1f),
+            green = (g.toFloat() / 255f).coerceIn(0f, 1f),
+            blue = (b.toFloat() / 255f).coerceIn(0f, 1f),
+            alpha = (a.toFloat() / 255f).coerceIn(0f, 1f)
         )
     }
